requirements.txt
flask
tensorflow
tensorflow-hub
numpy
pandas
Pillow
gunicorn
